+++ src/backend/db/migrations/index.ts
@@ -22,6 +22,7 @@ import { Add2FATotpSecret1587709969000 } from "./1587709969000-Add2FATotpSecret"
 import { DontCascadeSlotDeletion1588167544000 } from "./1588167544000-DontCascadeSlotDeletion";
 import { DeleteCascadePrefiledFor1591612977000 } from "./1591612977000-DeleteCascadePrefiledFor";
 import { SplitDisplayname1599542662000 } from "./1599542662000-SplitDisplayname";
+import { AddQuarantineReason1605087825000 } from "./1605087825000-AddQuarantineReason";
 
 export const migrations = [
   MakeSlotDateNullable1544015021618,
@@ -48,4 +49,5 @@ export const migrations = [
   DontCascadeSlotDeletion1588167544000,
   DeleteCascadePrefiledFor1591612977000,
   SplitDisplayname1599542662000,
+  AddQuarantineReason1605087825000,
 ];