+++ src/ui/entryReasonCategories.translation.ts
@@ -3,6 +3,7 @@ import { EntryReasonCategory } from "@@types";
 export const EntryReasonCategoriesTranslation = {
   en: {
     [EntryReasonCategory.ILLNESS]: "Illness / Doctor's Visit",
+    [EntryReasonCategory.QUARANTINE]: "Quarantine",
     [EntryReasonCategory.OTHER_EDUCATIONAL]: "Other (educational)",
     [EntryReasonCategory.OTHER_NON_EDUCATIONAL]: "Other (non-educational)",
     [EntryReasonCategory.EXAMEN]: "Examen",
@@ -13,6 +14,7 @@ export const EntryReasonCategoriesTranslation = {
   },
   de: {
     [EntryReasonCategory.ILLNESS]: "Krankheit / Arztbesuch",
+    [EntryReasonCategory.QUARANTINE]: "Quarantäne",
     [EntryReasonCategory.OTHER_EDUCATIONAL]: "Sonstiges (Schulisch)",
     [EntryReasonCategory.OTHER_NON_EDUCATIONAL]: "Sonstiges (Außerschulisch)",
     [EntryReasonCategory.EXAMEN]: "Klausur",